RF: Moved 'cohen_d' function from utils to processing module
